requirements.txt
pandas==2.1.0
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytest==7.4.0
httpx==0.25.0
scipy==1.11.3
scikit-surprise